+++ lib/clixon/clixon_datastore.h
@@ -73,7 +73,7 @@ cxobj *xmldb_cache_get(clicon_handle h, const char *db);
 
 int xmldb_modified_get(clicon_handle h, const char *db);
 int xmldb_modified_set(clicon_handle h, const char *db, int value);
-
+int xmldb_empty_get(clicon_handle h, const char *db);
 int xmldb_dump(clicon_handle h, FILE *f, cxobj *xt);
 
 #endif /* _CLIXON_DATASTORE_H */