+++ lib/src/clixon_xml.c
@@ -1883,6 +1883,7 @@ xml_copy_one(cxobj *x0,
     default:
 	break;
     }
+    xml_flag_set(x1, xml_flag(x0, XML_FLAG_DEFAULT)); /* Maybe more flags */
     retval = 0;
  done:
     return retval;