+++ test/test_nacm_default.sh
@@ -30,6 +30,7 @@ cat <<EOF > $cfg
   <CLICON_XMLDB_DIR>$dir</CLICON_XMLDB_DIR>
   <CLICON_RESTCONF_PRETTY>false</CLICON_RESTCONF_PRETTY>
   <CLICON_NACM_MODE>internal</CLICON_NACM_MODE>
+  <CLICON_NACM_DISABLED_ON_EMPTY>true</CLICON_NACM_DISABLED_ON_EMPTY>
   <CLICON_XMLDB_FORMAT>$format</CLICON_XMLDB_FORMAT>
 </clixon-config>
 EOF
@@ -62,7 +63,6 @@ EOF
 # 6: expected return value of test2
 # 7: expected return value of test3
 # 8: startup mode: startup or init
-# 9: Dont set default values (nullify them)
 testrun(){
     enablenacm=$1
     readdefault=$2
@@ -72,24 +72,9 @@ testrun(){
     ret2=$6
     ret3=$7
     db=$8
-    nulldef=$9
 
-    # Set default values (or not)
-    if [ $nulldef -ne 0 ]; then
-	# Defaults should be: true permit deny permit:
-	# nacm enabled, exec default permit, read permit (expect fail)"
-	# which means results should be 0 1 3
-	# Also enable-nacm is present since otherwise the nacm container would be removed
-	# since it is non-presence
-	NACM=$(cat <<EOF
-	       <nacm xmlns="urn:ietf:params:xml:ns:yang:ietf-netconf-acm">
-	            <enable-nacm>${enablenacm}</enable-nacm>
-               </nacm>
-EOF
-)
-    else
-	NACM=$(cat <<EOF
-	       <nacm xmlns="urn:ietf:params:xml:ns:yang:ietf-netconf-acm">
+    NACM=$(cat <<EOF
+    	       <nacm xmlns="urn:ietf:params:xml:ns:yang:ietf-netconf-acm">
 	            <enable-nacm>${enablenacm}</enable-nacm>
 		    <read-default>${readdefault}</read-default>
 		    <write-default>${writedefault}</write-default>
@@ -98,7 +83,6 @@ EOF
 	       </nacm>
 EOF
 )
-    fi
     # Initial data
     XML='<x xmlns="urn:example:nacm">42</x>'
 
@@ -117,6 +101,7 @@ EOF
 	start_backend -s $db -f $cfg
     else
 	new "Restart backend as eg follows: -Ff $cfg -s $db"
+	sleep 2
     fi
     
     new "waiting"
@@ -132,12 +117,16 @@ EOF
     wait_restconf
 
     # Use  POST (instead of startup)
+    # Note this only works because CLICON_NACM_DISABLED_ON_EMPTY is true
     if [ $db = init ]; then
-	new "Set Initial data using POST"
-	expectpart "$(curl -u guest:bar $CURLOPTS -X POST -H "Content-Type: application/yang-data+xml" -d "$XML" $RCPROTO://localhost/restconf/data)" 0 "HTTP/1.1 201 Created"
+	# Must set NACM first
+	new "Set NACM using PATCH"
+	expectpart "$(curl -u guest:bar $CURLOPTS -X PUT -H "Content-Type: application/yang-data+xml" -d "<data>$NACM$XML</data>" $RCPROTO://localhost/restconf/data)" 0 "HTTP/1.1 204 No Content"
+
+#	new "Set Initial data using POST"
+#	expectpart "$(curl -u guest:bar $CURLOPTS -X POST -H "Content-Type: application/yang-data+xml" -d "$XML" $RCPROTO://localhost/restconf/data)" 0 "HTTP/1.1 201 Created"
 	
-	new "Set NACM using POST"
-	expectpart "$(curl -u guest:bar $CURLOPTS -X POST -H "Content-Type: application/yang-data+xml" -d "$NACM" $RCPROTO://localhost/restconf/data)" 0 "HTTP/1.1 201 Created"
+
     fi
     
     #----------- First get
@@ -204,39 +193,36 @@ EOF
 # Run a lot of tests with different settings of default read/write/exec
 # Outer loop either starts from startup or inits config via restconf POST
 for db in startup init; do
-    new "nacm enabled and all defaults permit"
-    testrun true permit permit permit 0 0 0 $db 0
+    new "1. nacm enabled and all defaults permit"
+    testrun true permit permit permit 0 0 0 $db
 
-    new "nacm disabled and all defaults permit"
-    testrun false permit permit permit 0 0 0 $db 0
+    new "2. nacm disabled and all defaults permit"
+    testrun false permit permit permit 0 0 0 $db
 
-    new "nacm disabled and all defaults deny"
-    testrun false deny deny deny 0 0 0 $db 0
+    new "3. nacm disabled and all defaults deny"
+    testrun false deny deny deny 0 0 0 $db
 
-    new "nacm enabled, all defaults deny (expect fail)"
-    testrun true deny deny deny 1 1 1 $db 0
+    new "4. nacm enabled, all defaults deny (expect fail)"
+    testrun true deny deny deny 1 1 1 $db
 
-    new "nacm enabled, exec default deny - read permit (expect fail)"
-    testrun true permit deny deny 1 1 1 $db 0
+    new "5. nacm enabled, exec default deny - read permit (expect fail)"
+    testrun true permit deny deny 1 1 1 $db
 
-    new "nacm enabled, exec default deny - write permit (expect fail)"
-    testrun true deny permit deny 1 1 1 $db 0
+    new "6. nacm enabled, exec default deny - write permit (expect fail)"
+    testrun true deny permit deny 1 1 1 $db
 
-    new "nacm enabled, exec default deny read/write permit (expect fail)"
-    testrun true permit permit deny 1 1 1 $db 0
+    new "7. nacm enabled, exec default deny read/write permit (expect fail)"
+    testrun true permit permit deny 1 1 1 $db
 
-    new "nacm enabled, exec default permit, all others deny (expect fail)"
-    testrun true deny deny permit 2 1 2 $db 0
+    new "8. nacm enabled, exec default permit, all others deny (expect fail)"
+    testrun true deny deny permit 2 1 2 $db
 
-    new "nacm enabled, exec default permit, read permit (expect fail)"
-    testrun true permit deny permit 0 1 3 $db 0  # This is yang default
+    new "9. nacm enabled, exec default permit, read permit (expect fail)"
+    testrun true permit deny permit 0 1 3 $db  # This is yang default
 
-    new "nacm enabled, with default values (no settings - should be same as previous)"
-    # note last 1 means nullify all default values)
-    testrun true xxx xxx xxx 0 1 3 init 1 
+    new "10. nacm enabled, exec default permit, write permit (expect fail)"
+    testrun true deny permit permit 2 0 2 $db
 
-    new "nacm enabled, exec default permit, write permit (expect fail)"
-    testrun true deny permit permit 2 0 2 $db 0
 done
 
 rm -rf $dir